Refactored

+++ MyAnimeList/MALRequestClient.cs
@@ -3,20 +3,21 @@ using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
+using MyAnimeList.Exceptions;
 
 
 namespace MyAnimeList
 {
-    public class MALRequestClient
+    public static class MALRequestClient
     {
-        private static HttpClient _myAnimeListClient;
+        private static HttpClient? _myAnimeListClient;
         
         /// <summary>
         /// A static constructor used to create the httpclient
         /// </summary>
-        public static void Init()
+        public static async Task<bool> Init()
         {
-            //OAuth.Init();
+            await OAuth.Init();
             //OAuth.Login();
             
             if (_myAnimeListClient == null)
@@ -27,25 +28,31 @@ namespace MyAnimeList
                 _myAnimeListClient.DefaultRequestHeaders.Accept.Add(
                     new MediaTypeWithQualityHeaderValue("application/json"));
             }
+
+            return true;
         }
         
         /// <summary>
         /// Used to preform a get request to my anime list
         /// </summary>
-        /// <param name="url"> The url that a request will be made to </param>
-        /// <param name="requiresLogin"> Whether the request needs the user to be logged in to be fulfilled </param>
-        /// <typeparam name="T"> Generic return type </typeparam>
-        /// <returns> A formatted object of type T, null if bad request  </returns>
+        /// <param name="url">The url that a request will be made to</param>
+        /// <param name="requiresLogin">Whether the request needs the user to be logged in to be fulfilled</param>
+        /// <typeparam name="T">Generic return type</typeparam>
+        /// <returns>A formatted object of type T, null if bad request</returns>
         public static async Task<T?> Get<T>(string url, bool requiresLogin = false)
         {
+            if (_myAnimeListClient == null)
+            {
+                throw new HttpClientNotInitialized("Client not initialized, did you call the Init function?");
+            }
+            
             using var requestMessage = new HttpRequestMessage(HttpMethod.Get, url);
             if (requiresLogin)
-                requestMessage.Headers.Authorization = new AuthenticationHeaderValue("Bearer", "token goes here");
+                requestMessage.Headers.Authorization = new AuthenticationHeaderValue("Bearer", OAuth.GetToken());
             
             else
                 requestMessage.Headers.Add("X-MAL-CLIENT-ID", Constants.MyAnimeListConstants.ClientId);
             
-            /*Root myDeserializedClass = JsonSerializer.Deserialize<Root>(myJsonResponse)*/
             var response = await _myAnimeListClient.SendAsync(requestMessage);
             if (response.StatusCode == HttpStatusCode.OK)
             {
@@ -54,21 +61,34 @@ namespace MyAnimeList
             
             return default;
         }
-        
+
         /// <summary>
         /// Used to preform a get request to my anime list
         /// </summary>
         /// <param name="url"> The url that a request will be made to </param>
+        /// <param name="requiresLogin">Whether the request needs the user to be logged in to be fulfilled</param>
         /// <typeparam name="T"> Generic return type </typeparam>
         /// <returns> A formatted object of type T, null if bad request </returns>
-        public static async Task<T> Get<T>(Uri url)
+        public static async Task<T?> Get<T>(Uri url, bool requiresLogin = false)
         {
-            var response = await _myAnimeListClient.GetAsync(url);
+            if (_myAnimeListClient == null)
+            {
+                throw new HttpClientNotInitialized("Client not initialized, did you call the Init function?");
+            }
+            
+            using var requestMessage = new HttpRequestMessage(HttpMethod.Get, url);
+            if (requiresLogin)
+                requestMessage.Headers.Authorization = new AuthenticationHeaderValue("Bearer", "token goes here");
+            
+            else
+                requestMessage.Headers.Add("X-MAL-CLIENT-ID", Constants.MyAnimeListConstants.ClientId);
+            
+            var response = await _myAnimeListClient.SendAsync(requestMessage);
             if (response.StatusCode == HttpStatusCode.OK)
             {
                 return await response.Content.ReadAsAsync<T>();
             }
-
+            
             return default;
         }
         
@@ -80,6 +100,11 @@ namespace MyAnimeList
         /// <returns> True if the delete was successful else false </returns>
         public static async Task<bool> Delete(string url)
         {
+            if (_myAnimeListClient == null)
+            {
+                throw new HttpClientNotInitialized("Client not initialized, did you call the Init function?");
+            }
+            
             var response = await _myAnimeListClient.DeleteAsync(url);
             if (response.StatusCode == HttpStatusCode.OK)
             {
@@ -96,6 +121,11 @@ namespace MyAnimeList
         /// <returns> True if the delete was successful else false </returns>
         public static async Task<bool> Delete(Uri url)
         {
+            if (_myAnimeListClient == null)
+            {
+                throw new HttpClientNotInitialized("Client not initialized, did you call the Init function?");
+            }
+            
             var response = await _myAnimeListClient.DeleteAsync(url);
             if (response.StatusCode == HttpStatusCode.OK)
             {
@@ -112,8 +142,13 @@ namespace MyAnimeList
         /// <param name="body"> The request body </param>
         /// <typeparam name="T"> Generic return type </typeparam>
         /// <returns> A formatted object of type T, null if bad request </returns>
-        public static async Task<T> Patch<T>(string url, FormUrlEncodedContent body)
+        public static async Task<T?> Patch<T>(string url, FormUrlEncodedContent body)
         {
+            if (_myAnimeListClient == null)
+            {
+                throw new HttpClientNotInitialized("Client not initialized, did you call the Init function?");
+            }
+            
             var response = await _myAnimeListClient.PatchAsync(url, body);
             if (response.StatusCode == HttpStatusCode.OK)
             {
@@ -130,8 +165,13 @@ namespace MyAnimeList
         /// <param name="body"> The request body </param>
         /// <typeparam name="T"> Generic return type </typeparam>
         /// <returns> A formatted object of type T, null if bad request </returns>
-        public static async Task<T> Patch<T>(Uri url, FormUrlEncodedContent body)
+        public static async Task<T?> Patch<T>(Uri url, FormUrlEncodedContent body)
         {
+            if (_myAnimeListClient == null)
+            {
+                throw new HttpClientNotInitialized("Client not initialized, did you call the Init function?");
+            }
+            
             var response = await _myAnimeListClient.PatchAsync(url, body);
             if (response.StatusCode == HttpStatusCode.OK)
             {
